Model/requirements.txt
streamlit
ultralytics
numpy
Pillow
requests
opencv-python-headless